+++ src/Ops.Controllers/Areas/SiteManagement/SiteSettingsController.cs
@@ -1,11 +1,13 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Ocuda.Ops.Controllers.Abstract;
 using Ocuda.Ops.Controllers.Areas.SiteManagement.ViewModels.SiteSettings;
-using Ocuda.Ops.Models.Entities;
+using Ocuda.Ops.Service.Interfaces.Promenade.Services;
+using Ocuda.Promenade.Models.Entities;
 using Ocuda.Utility.Exceptions;
 using Ocuda.Utility.Keys;
 
@@ -16,19 +18,23 @@ namespace Ocuda.Ops.Controllers.Areas.SiteManagement
     [Route("[area]/[controller]")]
     public class SiteSettingsController : BaseController<SiteSettingsController>
     {
+        private readonly ISiteSettingPromService _siteSettingPromService;
+
         public static string Name { get { return "SiteSettings"; } }
         public static string Area { get { return "SiteManagement"; } }
 
-        public SiteSettingsController(ServiceFacades.Controller<SiteSettingsController> context)
-            : base(context)
+        public SiteSettingsController(ServiceFacades.Controller<SiteSettingsController> context,
+            ISiteSettingPromService siteSettingPromService) : base(context)
         {
+            _siteSettingPromService = siteSettingPromService 
+                ?? throw new ArgumentNullException(nameof(siteSettingPromService));
         }
 
         [Route("")]
         [Route("[action]")]
         public async Task<IActionResult> Index()
         {
-            var siteSettings = await _siteSettingService.GetAllAsync();
+            var siteSettings = await _siteSettingPromService.GetAllAsync();
             var categories = siteSettings.Select(_ => _.Category).Distinct();
             var siteSettingsByCategory = new Dictionary<string, List<SiteSetting>>();
 
@@ -64,7 +70,7 @@ namespace Ocuda.Ops.Controllers.Areas.SiteManagement
 
                     var key = model.SiteSetting.Id;
                     var value = model.SiteSetting.Value;
-                    var siteSetting = await _siteSettingService.UpdateAsync(key, value);
+                    var siteSetting = await _siteSettingPromService.UpdateAsync(key, value);
                     ShowAlertSuccess($"Updated {siteSetting.Name}");
                 }
                 catch (OcudaException ex)