+++ src/Ops.Controllers/Areas/SiteManagement/ViewModels/SocialCards/DetailViewModel.cs
@@ -1,0 +1,10 @@
+﻿using Ocuda.Promenade.Models.Entities;
+
+namespace Ocuda.Ops.Controllers.Areas.SiteManagement.ViewModels.SocialCards
+{
+    public class DetailViewModel
+    {
+        public SocialCard SocialCard { get; set; }
+        public string Action { get; set; }
+    }
+}