+++ src/Ops.Web/Startup.cs
@@ -212,6 +212,9 @@ namespace Ocuda.Ops.Web
                 Data.Promenade.EmediaCategoryRepository>();
             services.AddScoped<Service.Interfaces.Promenade.Repositories.IEmediaRepository,
                 Data.Promenade.EmediaRepository>();
+            services.AddScoped<
+                Service.Interfaces.Promenade.Repositories.IExternalResourcePromRepository,
+                Data.Promenade.ExternalResourcePromRepository>();
             services.AddScoped<Service.Interfaces.Promenade.Repositories.IFeatureRepository,
                 Data.Promenade.FeatureRepository>();
             services.AddScoped<Service.Interfaces.Promenade.Repositories.IGroupRepository,
@@ -239,6 +242,8 @@ namespace Ocuda.Ops.Web
                 Data.Promenade.SegmentRepository>();
             services.AddScoped<Service.Interfaces.Promenade.Repositories.ISegmentTextRepository,
                 Data.Promenade.SegmentTextRepository>();
+            services.AddScoped<Service.Interfaces.Promenade.Repositories.ISiteSettingPromRepository,
+                Data.Promenade.SiteSettingPromRepository>();
             services.AddScoped<Service.Interfaces.Promenade.Repositories.ISocialCardRepository,
                 Data.Promenade.SocialCardRepository>();
 
@@ -249,6 +254,7 @@ namespace Ocuda.Ops.Web
             services.AddScoped<ICategoryService, CategoryService>();
             services.AddScoped<IEmediaService, EmediaService>();
             services.AddScoped<IEmailService, EmailService>();
+            services.AddScoped<IExternalResourcePromService, ExternalResourcePromService>();
             services.AddScoped<IExternalResourceService, ExternalResourceService>();
             services.AddScoped<IFeatureService, FeatureService>();
             services.AddScoped<IFileService, FileService>();
@@ -272,6 +278,7 @@ namespace Ocuda.Ops.Web
             services.AddScoped<IScheduleRequestService, ScheduleRequestService>();
             services.AddScoped<ISectionService, SectionService>();
             services.AddScoped<ISegmentService, SegmentService>();
+            services.AddScoped<ISiteSettingPromService, SiteSettingPromService>();
             services.AddScoped<ISiteSettingService, SiteSettingService>();
             services.AddScoped<ISocialCardService, SocialCardService>();
             services.AddScoped<Service.Abstract.IUserContextProvider, UserContextProvider>();